chore: ignore changes to .DS_Store file ```
Since `.DS_Store` is a macOS system file that stores custom attributes of folders (like icon positions) and is not relevant to the codebase, it's common practice to ignore changes to this file. This commit doesn't introduce any functional changes but rather acknowledges the presence of this system file. The type "chore" is appropriate as it's a maintenance task that doesn't fit into other categories.